Refactor packages
Packages should imply something more about the classes contained
within. To be fair, I think these package names are only marginally less
confusing that their predecessors, but I couldn't think of anything
better :)
