+++ src/main/java/net/imglib2/imagej/imageplus/package-info.java
@@ -1,0 +1,5 @@
+package net.imglib2.imagej.imageplus;
+
+/**
+ * This package contains utilities for converting from {@link ij.ImagePlus} to {@link net.imglib2.RandomAccessibleInterval} and its various subtypes.
+ */
