+++ src/main/java/net/imglib2/imagej/img/AbstractVirtualStack.java
@@ -32,7 +32,7 @@
  * #L%
  */
 
-package net.imglib2.imagej.display;
+package net.imglib2.imagej.img;
 
 import java.awt.Rectangle;
 import java.awt.image.ColorModel;
