+++ src/main/java/net/imglib2/imagej/img/CalibrationUtils.java
@@ -31,7 +31,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imglib2.imagej.display;
+package net.imglib2.imagej.img;
 
 import java.util.ArrayList;
 import java.util.List;
