+++ src/main/java/net/imglib2/imagej/img/ImageProcessorUtils.java
@@ -32,7 +32,7 @@
  * #L%
  */
 
-package net.imglib2.imagej.display;
+package net.imglib2.imagej.img;
 
 import ij.process.ByteProcessor;
 import ij.process.ColorProcessor;
