+++ src/main/java/net/imglib2/imagej/img/package-info.java
@@ -1,0 +1,5 @@
+package net.imglib2.imagej.img;
+
+/**
+ * This package contains utilities for converting from {@link net.imglib2.RandomAccessibleInterval} and its various subtypes to {@link ij.ImagePlus}.
+ */
