+++ src/test/java/net/imglib2/imagej/img/CellImgToImagePlusTest.java
@@ -32,7 +32,7 @@
  * #L%
  */
 
-package net.imglib2.imagej.display;
+package net.imglib2.imagej.img;
 
 import ij.ImagePlus;
 import ij.ImageStack;
